python/flaskApi/requirementsfordocker.txt
Flask==1.1.2
gevent==21.1.2
gunicorn==20.1.0
numpy==1.19.5
opencv-python-headless
Pillow==8.2.0
